+++ CHANGELOG.md
@@ -7,8 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 ### Removed
 ### Changed
+- The utility function `dhtSync` parameter `timeoutMs` now defaults to 60000 milliseconds, and the parameter `intervalMs` now defaults to 500 milliseconds.
+
 ### Fixed
 - Return an error if spawning `kitsune2-bootstrap-srv` fails
+- The utility functions `dhtSync` and `conductorCellsDhtSync` now check that all Ops have been integrated and are not in limbo.
 
 ## 2025-04-11: v0.18.0-rc.1
 ### Fixed